Implement drag&drop (bug 620213)
Allows to drop files from, e.g., a file manager onto the image list.
If the file extension is ".pdf", the file is interpreted as pdf.
Otherwise it is attempted to be loaded as an image.

Note: No elaborate mime-type checking is performed.

Most of the code is adapted from:
http://faq.pygtk.org/index.py?req=show&file=faq23.031.htp