Changes the way the editor of the content areas is implemented
These are big changes that remove the usage of a GtkNotebook that
stores every editor instance and instead use just one instance to
handle the the currently selected area.

With these changes, the settings of the line and letter spacing of
an area are lost when selecting a different area. This is supposed
to be fixed later.