Change the async worker to use threads for its job execution
The async worker is only one thread to decouple the job execution from
the ui. This introduces per-job-threads within the async worker. This
allows to execute more than one job at a time in the worker. The maximum
spawned threads are controlled by a semaphore. The count is hardcoded at
the moment.

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>