add a multiprocess Pool to the async worker.
The multiprocess.Pool is a Process pool for doing parallel work without
a limiting GIL - so its really parallel if you have more than one cpu.

The Pool usage is limited for cases where the "parallel" argument of the
AsyncWorker is >1 - in other cases this makes no sense and is uneccesary
overhead.

Another benefit is, that there are less shared-state-problems that
treads normally introduce - because of the real process insolation.

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>