fix a race condition in parallel recognizion
The worker had a state to signal that he has done. This state was set to
late so that no finish callback see that the worker is ready.

This commit simply drop the extra state and use the counter counting how
many callbacks where called. If the counter has the same value as the
page count it assumes that this is the last one - and updates, close the
dialog, ...

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>