Prevent exceptions when setting scan mode and resolution
Invalid scanner options lead to an unrecoverable error:
Error scanning page

We want to set the mode as color and resolution as 300 DPI if possible,
but the setting unsupported values was resulting in an non-caught
exception.

Some scanners don't support the Resolution parameter at all.

Supported modes may be vary across different scanners.
Not all of them support the "Color" mode.
For example, some of them support "Color24" or "Color48" instead.

This patch fixes that by simply catching the exceptions, using the
default values in that case.