Do not ship odf package
A long, long time ago, when OCRFeeder was first released, there was
no Debian package for odfpy, so to it simple at the time it was shipped
with the rest of the source code.

There are packages available now, so we should drop odf from the source
tree, allowing to use updated versions with bug and security fixes.

This will also make it easier to port OCRFeeder to Python 3 as the odfpy
package doesn't need to be included in the port.