Fix setting the OCR engine for the first time in the BoxEditor
The code that sets the OCR engine for the BoxEditor was not getting
the current_engine set the first time it checks it, so it was not
trying to get the index for that engine, and defaulted it to None.

Since it was later comparing it to an int, there was a TypeError
exception because of that.