Port to Python 3
Python 3 has been around for a while now, and Python 2 is going away
in Debian soon, so a port was due.

These changes represent the port and were done by using lib2to3, with
a few manual editions to add more needed changes (like dropping the
attempt to encode a string), or dropping some unnecessary ones, e.g.
iterating through an object as returned by the dict's keys() method
is fine, but lib2to3 still creates a list for it.

Special thanks to Reuben (@reubot on GNOME Gitlab) for having done an
initial MR for the port, which made me want to move this foward (but
I am pushing a different commit as it needed many changes).