Handle special chars in filenames when importing PDF files
Some special characters (e.g. quotes) in the filename cause gs to fail.
What's worse, gs interprets the escape character as a real character.
This means that it cannot handle all Unix files on its own.
We need to create a temp symlink as a workaround for gs limitations.

Fixes GNOME/ocrfeeder#20