util: Change how ISO_CODES_PATH is set
The ISO_CODES_PATH was hardcoded, and this means that the path may not
be the correct one in some systems. Therefore, these changes allow to
set an ISO_CODES_DIR as an environment variable or, in case it is not
set, it looks for the "xml/iso-codes" path within the system's data
dirs.